FDG-10264 create legend
